requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
attrs==20.3.0             # via jsonschema
django==3.1.3             # via -r requirements.in, djangorestframework, djangorestframework-recursive, drf-spectacular
djangorestframework-recursive==0.1.2  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, djangorestframework-recursive, drf-spectacular
drf-spectacular==0.11.1   # via -r requirements.in
inflection==0.5.1         # via drf-spectacular
jsonschema==3.2.0         # via drf-spectacular
pyrsistent==0.17.3        # via jsonschema
pytz==2020.4              # via django
pyyaml==5.3.1             # via drf-spectacular
six==1.15.0               # via jsonschema
sqlparse==0.4.1           # via django
uritemplate==3.0.1        # via drf-spectacular

# The following packages are considered to be unsafe in a requirements file:
# setuptools
